iris_offline/requirements.txt
# IRIS 2.0 Offline Module — Python Requirements
# Install via: pip install -r requirements.txt
# Or use setup.sh for full automated setup.

# ─── Core inference ──────────────────────────────────────────────────────────
ultralytics>=8.0.200         # YOLOv8n — installs torch CPU automatically
numpy>=1.24,<2.0             # Pinned <2.0 for ultralytics compat

# ─── Computer vision ────────────────────────────────────────────────────────
# Use headless build on Raspberry Pi (no Qt dependency)
opencv-python-headless>=4.8

# ─── Web server ─────────────────────────────────────────────────────────────
flask>=3.0
werkzeug>=3.0
pyopenssl>=23.0              # HTTPS SSL context for Flask
cryptography>=41.0           # Required by pyopenssl

# ─── Hardware I/O ────────────────────────────────────────────────────────────
smbus2>=0.4                  # I2C for MPU9250
RPi.GPIO>=0.7.1              # GPIO for HC-SR04 (RPi only)
